refactor: split auth route handlers
