+++ src/stdlib/auth.rs
@@ -45,6 +45,7 @@ use totp_rs::{Algorithm as TotpAlgorithm, Secret, TOTP};
 
 mod cookies;
 mod providers;
+mod routes;
 
 pub use cookies::default_auth_cookie_secure_env;
 use cookies::{
@@ -56,6 +57,10 @@ use providers::{
     validate_provider_name, value_map_to_provider,
 };
 pub use providers::{provider_to_value, value_to_provider};
+pub use routes::{
+    handle_auth_callback, handle_auth_index, handle_auth_logout, handle_auth_protect,
+    handle_auth_start,
+};
 
 // Internal architecture note:
 // Phase 4.5 is being executed in small, behavior-preserving slices. `auth.rs` remains the
@@ -5183,437 +5188,6 @@ fn json_response(data: Value, status: i64) -> Value {
     response.insert("body".to_string(), Value::String(body));
 
     Value::Map(response)
-}
-
-pub fn handle_auth_start(args: &[Value]) -> Result<Value> {
-    let req = &args[0];
-
-    // Get provider name from route params
-    let provider_name = if let Value::Map(req_map) = req {
-        req_map.get("params").and_then(|p| {
-            if let Value::Map(params) = p {
-                params.get("provider").and_then(|v| {
-                    if let Value::String(s) = v {
-                        Some(s.clone())
-                    } else {
-                        None
-                    }
-                })
-            } else {
-                None
-            }
-        })
-    } else {
-        None
-    };
-
-    let provider_name = provider_name.ok_or_else(|| {
-        IntentError::runtime_error("[auth] No provider specified in /auth/{provider}".to_string())
-    })?;
-
-    // Get auth config
-    let config = get_auth_config().ok_or_else(|| {
-        IntentError::runtime_error(
-            "[auth] Auth not configured - call enable_auth() first".to_string(),
-        )
-    })?;
-
-    // Find the provider
-    let provider = config
-        .providers
-        .iter()
-        .find(|p| p.name == provider_name)
-        .ok_or_else(|| {
-            let msg = if let Some(suggestion) = suggest_provider(&provider_name) {
-                format!(
-                    "[auth] Unknown provider \"{}\"\n       Did you mean \"{}\"?\n       Available providers: {}",
-                    provider_name, suggestion, available_providers()
-                )
-            } else {
-                format!(
-                    "[auth] Unknown provider \"{}\"\n       Available providers: {}",
-                    provider_name, available_providers()
-                )
-            };
-            IntentError::runtime_error(msg)
-        })?;
-
-    // Generate state for CSRF protection
-    let state = generate_oauth_state();
-
-    // Generate nonce for OIDC
-    let nonce = if provider.supports_oidc {
-        Some(generate_nonce())
-    } else {
-        None
-    };
-
-    // Generate PKCE if enabled
-    let (pkce_verifier, pkce_challenge) = if provider.use_pkce {
-        let verifier = generate_pkce_verifier();
-        let challenge = generate_pkce_challenge(&verifier);
-        (Some(verifier), Some(challenge))
-    } else {
-        (None, None)
-    };
-
-    // Determine redirect URI (provider-specific for GitHub compatibility)
-    let (host, proto) = get_host_and_proto(req);
-    let redirect_uri = format!("{}://{}/auth/{}/callback", proto, host, provider.name);
-
-    // Store OAuth state
-    store_oauth_state(
-        &state,
-        &provider.name,
-        &redirect_uri,
-        nonce.as_deref(),
-        pkce_verifier.as_deref(),
-    );
-
-    // When store_tokens is enabled and provider is Google, ensure we request refresh tokens
-    let mut provider_for_url = provider.clone();
-    if config.store_tokens {
-        if !provider_for_url.extra_params.contains_key("access_type") {
-            if provider_for_url.authorize_url.contains("google") {
-                provider_for_url
-                    .extra_params
-                    .insert("access_type".to_string(), "offline".to_string());
-                // prompt=consent forces Google to return a new refresh token every time
-                if !provider_for_url.extra_params.contains_key("prompt") {
-                    provider_for_url
-                        .extra_params
-                        .insert("prompt".to_string(), "consent".to_string());
-                }
-            }
-        }
-    }
-
-    // Generate auth URL
-    let auth_url = generate_auth_url(
-        &provider_for_url,
-        &redirect_uri,
-        &state,
-        nonce.as_deref(),
-        pkce_challenge.as_deref(),
-    );
-
-    Ok(redirect_response(&auth_url, None))
-}
-
-/// Auth middleware entry point used for protected route enforcement.
-pub fn handle_auth_protect(args: &[Value]) -> Result<Value> {
-    match enforce_auth_for_request(&args[0], false) {
-        Ok(()) => Ok(Value::Unit),
-        Err(response) => Ok(response),
-    }
-}
-
-/// Handle GET /auth - minimal auth landing route.
-///
-/// If exactly one provider is configured, redirect straight into that OAuth flow.
-/// If multiple providers are configured, render a small provider chooser.
-pub fn handle_auth_index(_args: &[Value]) -> Result<Value> {
-    let config = get_auth_config()
-        .ok_or_else(|| IntentError::runtime_error("[auth] Auth not configured".to_string()))?;
-
-    if config.providers.len() == 1 {
-        let provider = &config.providers[0];
-        let safe_provider = encode_url_path_segment(&provider.name);
-        return Ok(redirect_response(&format!("/auth/{}", safe_provider), None));
-    }
-
-    let provider_links = config
-        .providers
-        .iter()
-        .map(|provider| {
-            let label = escape_html(&provider.name);
-            let safe_provider = encode_url_path_segment(&provider.name);
-            format!(
-                r#"<li><a href="/auth/{name}">Sign in with {label}</a></li>"#,
-                name = safe_provider,
-                label = label
-            )
-        })
-        .collect::<Vec<_>>()
-        .join(
-            "
-",
-        );
-
-    Ok(html_response(&format!(
-        r#"<!doctype html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Sign in</title>
-  </head>
-  <body>
-    <main>
-      <h1>Sign in</h1>
-      <p>Choose a provider to continue.</p>
-      <ul>
-        {provider_links}
-      </ul>
-    </main>
-  </body>
-</html>"#,
-    )))
-}
-
-/// Handle GET /auth/callback - OAuth callback
-///
-/// Uses a two-phase flow to work around Safari ITP cookie restrictions.
-/// Phase 1: Receives OAuth code, creates session, returns intermediate page.
-/// Phase 2: Intermediate page redirects back with exchange token, cookie is set.
-pub fn handle_auth_callback(args: &[Value]) -> Result<Value> {
-    let req = &args[0];
-
-    let config = get_auth_config()
-        .ok_or_else(|| IntentError::runtime_error("[auth] Auth not configured".to_string()))?;
-
-    // Extract code, state, error, and exchange from query params
-    let (code, state, error, exchange) = if let Value::Map(req_map) = req {
-        if let Some(Value::Map(query)) = req_map.get("query_params") {
-            let code = query.get("code").and_then(|v| {
-                if let Value::String(s) = v {
-                    Some(s.clone())
-                } else {
-                    None
-                }
-            });
-            let state = query.get("state").and_then(|v| {
-                if let Value::String(s) = v {
-                    Some(s.clone())
-                } else {
-                    None
-                }
-            });
-            let error = query.get("error").and_then(|v| {
-                if let Value::String(s) = v {
-                    Some(s.clone())
-                } else {
-                    None
-                }
-            });
-            let exchange = query.get("exchange").and_then(|v| {
-                if let Value::String(s) = v {
-                    Some(s.clone())
-                } else {
-                    None
-                }
-            });
-            (code, state, error, exchange)
-        } else {
-            (None, None, None, None)
-        }
-    } else {
-        (None, None, None, None)
-    };
-
-    // Check for OAuth error
-    if let Some(err) = error {
-        eprintln!("[auth] OAuth error: {}", err);
-        return Ok(redirect_response(&config.failure_url, None));
-    }
-
-    // Phase 2: Exchange token flow (same-origin redirect from intermediate page)
-    // This breaks the cross-site redirect chain so Safari ITP doesn't cap cookie lifetime.
-    if let Some(exchange_token) = exchange {
-        if code.is_some() {
-            // Both `exchange` and `code` in the same request is not a legitimate flow.
-            // Consume the exchange token to prevent it lingering, then fall through
-            // to Phase 1 (which will fail on missing state — correct behavior).
-            let _ = consume_exchange_token(&exchange_token);
-            eprintln!(
-                "[auth] Ignoring exchange token — OAuth code also present (crafted request?)"
-            );
-        } else {
-            // Consume the exchange token to get the session ID
-            if let Some(session_id) = consume_exchange_token(&exchange_token) {
-                // Verify the session still exists
-                if let Some(_session) = get_session_by_id(&session_id) {
-                    eprintln!(
-                        "[auth] Session exchange token consumed for session {}...",
-                        &session_id[..8]
-                    );
-
-                    let cookie = build_signed_session_cookie(&config, &session_id, None)
-                        .map_err(IntentError::runtime_error)?;
-
-                    return Ok(redirect_response(&config.success_url, Some(&cookie)));
-                } else {
-                    eprintln!("[auth] Exchange token valid but session not found");
-                    return Ok(redirect_response(&config.failure_url, None));
-                }
-            } else {
-                eprintln!("[auth] Invalid or expired exchange token");
-                return Ok(redirect_response(&config.failure_url, None));
-            }
-        }
-    }
-
-    // Phase 1: Standard OAuth callback flow
-    // Validate state
-    let oauth_state = state.as_ref().and_then(|s| consume_oauth_state(s));
-
-    if oauth_state.is_none() || code.is_none() {
-        eprintln!("[auth] Invalid callback - missing code or state");
-        return Ok(redirect_response(&config.failure_url, None));
-    }
-
-    let oauth_state = oauth_state.unwrap();
-    let code = code.unwrap();
-
-    // Find the provider
-    let provider = config
-        .providers
-        .iter()
-        .find(|p| p.name == oauth_state.provider);
-    if provider.is_none() {
-        eprintln!("[auth] Provider not found: {}", oauth_state.provider);
-        return Ok(redirect_response(&config.failure_url, None));
-    }
-    let provider = provider.unwrap();
-    let provider_name = provider.name.clone();
-
-    // Exchange code for tokens
-    let tokens = match exchange_code_for_tokens(
-        provider,
-        &code,
-        &oauth_state.redirect_url,
-        oauth_state.pkce_verifier.as_deref(),
-    ) {
-        Ok(t) => t,
-        Err(e) => {
-            eprintln!("{}", e);
-            return Ok(redirect_response(&config.failure_url, None));
-        }
-    };
-
-    // Get user info - prefer ID token for OIDC providers
-    let user_info = if let Some(ref id_token) = tokens.id_token {
-        // Decode and validate ID token
-        match decode_id_token(id_token) {
-            Ok(claims) => {
-                // Validate ID token claims
-                if let Err(e) = validate_id_token_claims(
-                    &claims,
-                    provider.issuer.as_deref(),
-                    &provider.client_id,
-                    oauth_state.nonce.as_deref(),
-                ) {
-                    eprintln!("{}", e);
-                    return Ok(redirect_response(&config.failure_url, None));
-                }
-                claims
-            }
-            Err(e) => {
-                eprintln!(
-                    "[auth] ID token decode failed, falling back to userinfo: {}",
-                    e
-                );
-                // Fall back to userinfo endpoint
-                match fetch_userinfo(provider, &tokens.access_token) {
-                    Ok(u) => u,
-                    Err(e) => {
-                        eprintln!("{}", e);
-                        return Ok(redirect_response(&config.failure_url, None));
-                    }
-                }
-            }
-        }
-    } else {
-        // No ID token, use userinfo endpoint
-        match fetch_userinfo(provider, &tokens.access_token) {
-            Ok(u) => u,
-            Err(e) => {
-                eprintln!("{}", e);
-                return Ok(redirect_response(&config.failure_url, None));
-            }
-        }
-    };
-
-    // Create session (optionally storing tokens)
-    let session = create_session(
-        &provider_name,
-        user_info,
-        if config.store_tokens {
-            Some(&tokens)
-        } else {
-            None
-        },
-        config.session_ttl,
-    )
-    .map_err(|e| IntentError::runtime_error(format!("[auth] Failed to create session: {}", e)))?;
-    let mut session = session;
-    if let Some(existing_session_id) = get_session_id_from_request(req) {
-        if let Some(existing_session) = get_session_by_id(&existing_session_id) {
-            if existing_session_id != session.id {
-                if session.data_json == "{}" && existing_session.data_json != "{}" {
-                    session.data_json = existing_session.data_json.clone();
-                }
-                migrate_session(&existing_session_id, &session).map_err(|e| {
-                    IntentError::runtime_error(format!(
-                        "[auth] Failed to rotate session after OAuth callback: {}",
-                        e
-                    ))
-                })?;
-            } else {
-                store_session(session.clone());
-            }
-        } else {
-            store_session(session.clone());
-        }
-    } else {
-        store_session(session.clone());
-    }
-    let session_id = session.id.clone();
-
-    // Generate exchange token and store it
-    let exchange_token = generate_session_id();
-    store_exchange_token(&exchange_token, &session_id);
-
-    // Return intermediate HTML page that does a same-origin redirect.
-    // This breaks the cross-site redirect chain so Safari ITP treats the
-    // subsequent Set-Cookie as first-party (full lifetime instead of ~1 hour).
-    // Uses meta-refresh as primary (works without JS, unaffected by CSP),
-    // JS redirect as fast-path, and a clickable link as final fallback.
-    //
-    // URL-encode provider name defensively — provider names are typically
-    // alphanumeric (e.g. "google") but encoding prevents JS string breakage
-    // if a name ever contains quotes or special chars. Exchange token is a UUID.
-    let safe_provider = encode_url_path_segment(&provider_name);
-    let callback_url = format!(
-        "/auth/{}/callback?exchange={}",
-        safe_provider, exchange_token
-    );
-    let html = format!(
-        r#"<!DOCTYPE html>
-<html><head><meta charset="utf-8"><title>Completing login...</title>
-<meta http-equiv="refresh" content="0;url={url}"></head>
-<body><p>Completing login...</p>
-<script>window.location.replace("{url}");</script>
-<noscript><a href="{url}">Click here to continue</a></noscript>
-</body></html>"#,
-        url = callback_url
-    );
-
-    Ok(html_response(&html))
-}
-
-/// Handle POST /auth/logout - Clear session
-pub fn handle_auth_logout(args: &[Value]) -> Result<Value> {
-    let req = &args[0];
-    let config = get_auth_config().unwrap_or_default();
-
-    if let Some(session_id) = get_session_id_from_request(req) {
-        delete_session_by_id(&session_id);
-    }
-
-    let cookie = build_cleared_session_cookie(&config, None).map_err(IntentError::runtime_error)?;
-
-    Ok(redirect_response(&config.logout_url, Some(&cookie)))
 }
 
 // ============================================================================
